Create README.md
This program creates three arrays: car Models, carYears and carPrices, which contain information about several cars. Then it displays a list of cars and reads the car number from the keyboard. If a car with the specified number exists, the program displays information about the car. Otherwise, the program displays an error message.

You can run the program and enter different car numbers to check how it will work in different situations.